Move files a bit to order the test code
