+++ test/lib/json-diff.ts
