+++ test/lib/util.ts
@@ -1,3 +1,4 @@
+import compactStringify from 'json-stringify-pretty-compact';
 import {ICanonicalTileID, ILngLatLike} from '../../src';
 import {Point2D} from '../../src/point2d';
 import Color from '../../src/util/color';
@@ -54,4 +55,18 @@ export function getPoint(coord: ILngLatLike, canonical: ICanonicalTileID): Point
     p.x = Math.round(p.x);
     p.y = Math.round(p.y);
     return p;
+}
+
+// we have to handle this edge case here because we have test fixtures for this
+// edge case, and we don't want UPDATE=1 to mess with them
+export function stringify(v) {
+    let s = compactStringify(v);
+
+    if (s.indexOf('\u2028') >= 0) {
+        s = s.replace(/\u2028/g, '\\u2028');
+    }
+    if (s.indexOf('\u2029') >= 0) {
+        s = s.replace(/\u2029/g, '\\u2029');
+    }
+    return s;
 }
